game_over_added

+++ Source/config.h
@@ -3,8 +3,8 @@
 
 class Config{
 public:
-    int multiplier = 1;
-    int threshold = 4;
+    double multiplier = 1;
+    double threshold = 4;
 };
 
 #endif // CONFIG_H
